algorithmic_sciences_task/requirements.txt
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.2
coverage==6.0.1
cycler==0.12.1
docformatter==1.7.5
flake8==7.2.0
flake8-docstrings==1.7.0
fonttools==4.57.0
iniconfig==2.1.0
kiwisolver==1.4.8
matplotlib==3.10.1
mccabe==0.7.0
mypy==1.15.0
mypy_extensions==1.1.0
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pluggy==1.5.0
psutil==7.0.0
pycodestyle==2.13.0
pydocstyle==6.3.0
pyflakes==3.3.2
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
seaborn==0.13.2
six==1.17.0
snowballstemmer==2.2.0
types-PyYAML==6.0.12.20250402
typing_extensions==4.13.2
tzdata==2025.2
untokenize==0.1.1
